Clarify dividends requirements
Both "pari passu" and "pro rate" are archaic Latin and vague. This patch
replace both, which appeared in a single sentence, with more explicit
language. The goal is to prevent dividends that preference specific
holders or securities over others. Dividends disproportionate to shares
held are a reliable red flag.

Per _A Dictionary of Model Legal Usage_ (second edition):

> pari passu (= with equal pace; equally; at the same time) ...

> pro rata, adv., .... Proportionately will sometimes serve in place of
> pro rata.

The problematic history of "pari passu" is covered at book length in
Gulati and Scott's _The Three and a Half Minute Transaction_.